Removed a broken link from "index" to "contents" page.
The "contents" page contains the global ToC tree, but actually it is not
needed as a separate page.  Unfortunately, there is no easy way to avoid
rendering of this page in Sphinx.  As a workaround, this page is skipped
during deployment.  But Sphinx still includes it in the global ToC and
generates a "prev" link to it in the "index" page.

Alternative solution could be removing the "contents" page and putting
global ToC into the "index" page.  But there is a problem with self
referencing in Sphinx: https://github.com/sphinx-doc/sphinx/issues/2103

So, as a workaround, now the layout is altered to explicitly avoid "prev"
link from the "index" page.